Move fake repos to integration test fodler
