+++ critique_wheel/tests/unit/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-import tests.e2e.fake_iam_repository as fake_iam_repository
+import tests.integration.fake_iam_repository as fake_iam_repository
 from critique_wheel.members.services.iam_service import IAMService
 
 
@@ -12,13 +12,3 @@ def iam_repo():
 @pytest.fixture
 def iam_service(iam_repo):
     return IAMService(iam_repo)
-
-
-# @pytest.fixture
-# def work_repo():
-#     return fake_work_repository.FakeWorkRepository([])
-
-
-# @pytest.fixture
-# def work_service(work_repo):
-#     return WorkService(work_repo)
